fix: do try to delete api in management if no id is present

+++ controllers/apim/apidefinition/internal/delete.go
@@ -18,7 +18,7 @@ func (d *Delegate) Delete(
 	}
 
 	var apiNotFoundError *managementapierror.ApiNotFoundError
-	if d.IsConnectedToManagementApi() {
+	if d.IsConnectedToManagementApi() && apiDefinition.Status.ID != "" {
 		d.log.Info("Delete API definition into Management API")
 		err := d.apimClient.DeleteApi(apiDefinition.Status.ID)
 		if errors.As(err, &apiNotFoundError) {
